# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TerraBovina API
  version: "2.0.0"
  description: |
    API REST de TerraBovina (V2): gestión de estancias, potreros y ganado con
    geometría espacial (MySQL 8 / ST_*), asignación y traslado de animales
    entre potreros, ingesta de NDVI/clima y estimaciones forrajera/nutricional
    respaldadas por un modelo predictivo externo, y recomendaciones de manejo.

    Todos los endpoints bajo `/api/v2` requieren `Authorization: Bearer <token>`
    excepto `/auth/*`. Un recurso ajeno o inexistente siempre responde **404**
    (nunca 403), para no revelar su existencia a otro usuario.
  contact:
    name: TerraBovina
  license:
    name: ISC
    url: https://opensource.org/license/isc-license-txt
servers:
  - url: https://backendterrabovina-production.up.railway.app/api/v2
    description: Producción
  - url: http://localhost:3000/api/v2
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /potrero/{potreroId}/estimacion-forrajera:
    parameters:
      - $ref: "#/components/parameters/potreroId"
    post:
      tags: [Estimaciones]
      operationId: crearEstimacionForrajera
      summary: Generar disponibilidad forrajera (modelo predictivo real)
      description: >-
        Llama a `POST /predict/dmp` del modelo predictivo externo (imagen
        satelital + clima + predicción, todo en una respuesta) y persiste
        ObservacionSatelital + DatoClimatico + DisponibilidadForrajera en una
        transacción. Si el modelo no responde, no se persiste nada: la última
        fila sigue siendo la vigente para los GET de NDVI/clima.
      responses:
        "201":
          description: Disponibilidad forrajera calculada y persistida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EstimacionForrajeraResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502":
          description: El modelo predictivo no respondió o respondió con error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /potrero/{potreroId}/estimacion-nutricional:
    parameters:
      - $ref: "#/components/parameters/potreroId"
    post:
      tags: [Estimaciones]
      operationId: crearEstimacionNutricional
      summary: Generar demanda nutricional (modelo predictivo real)
      description: >-
        Arma el lote de animales con asignación vigente en el potrero, llama
        a `POST /predict/dmi` del modelo predictivo externo (predicción por
        animal) y persiste el agregado (suma de `dmi_kg_dia`) como una
        EstimacionDemanda del potrero. `advertencias` y `predicciones` por
        animal viajan en la respuesta pero no se persisten.
      responses:
        "201":
          description: Demanda nutricional calculada y persistida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EstimacionNutricionalResponse" }
        "400":
          description: El potrero no tiene animales con asignación vigente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502":
          description: El modelo predictivo no respondió o respondió con error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Token JWT devuelto por POST /auth/login, firmado con id_usuario y
        email en el payload.
  parameters:
...
    potreroId:
      name: potreroId
      in: path
      required: true
      schema: { type: integer }
...
  responses:
...
    Unauthorized:
      description: Falta el token de sesión, o es inválido/venció
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: >-
        El recurso no existe, o pertenece a otro usuario (mismo código en
        ambos casos, para no revelar la existencia de recursos ajenos).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ObservacionSatelital:
      type: object
      properties:
        id_observacion: { type: integer }
        id_potrero: { type: integer }
        fuente: { type: string, example: SENTINEL2 }
        fecha: { type: string, format: date }
        ndvi: { type: string, nullable: true, description: "DECIMAL(6,4) serializado como string" }
        nubosidad: { type: string, nullable: true, description: "DECIMAL(5,2) serializado como string" }
        created_at: { type: string, format: date-time }
    DatoClimatico:
      type: object
      properties:
        id_dato: { type: integer }
        id_potrero: { type: integer }
        fuente: { type: string, example: MODELO_PREDICTIVO }
        fecha: { type: string, format: date }
        temperatura: { type: string, nullable: true, description: "DECIMAL(5,2) serializado como string" }
        precipitacion: { type: string, nullable: true, description: "DECIMAL(6,2) serializado como string" }
        humedad: { type: string, nullable: true, description: "DECIMAL(5,2) serializado como string" }
        created_at: { type: string, format: date-time }
    DisponibilidadForrajera:
      type: object
      properties:
        id_disponibilidad: { type: integer }
        id_potrero: { type: integer }
        fecha_calculo: { type: string, format: date-time }
        kg_materia_seca_ha: { type: string, description: "DECIMAL(8,2) serializado como string" }
        superficie_analizada_ha: { type: string, nullable: true }
        indice_ndvi: { type: string, nullable: true, description: "DECIMAL(6,4) serializado como string" }
        version_modelo: { type: string, example: dmp-model }
        nivel_confianza: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    EstimacionForrajeraResponse:
      allOf:
        - $ref: "#/components/schemas/DisponibilidadForrajera"
        - type: object
          properties:
            dmp_kg_ms_ha_periodo:
              type: number
              format: double
              description: Predicción del modelo para la ventana del período (no se persiste todavía)
            observacion: { $ref: "#/components/schemas/ObservacionSatelital" }
            clima: { $ref: "#/components/schemas/DatoClimatico" }
    EstimacionDemanda:
      type: object
      properties:
        id_estimacion: { type: integer }
        id_potrero: { type: integer }
        fecha_calculo: { type: string, format: date-time }
        cantidad_animales: { type: integer }
        kg_materia_seca_dia: { type: string, description: "DECIMAL(8,2) serializado como string" }
        version_modelo: { type: string, example: dmi-model }
        nivel_confianza: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    DmiPrediccionAnimal:
      type: object
      properties:
        animal_id: { type: string, description: Se envía como String(id_ganado) }
        dmi_kg_dia: { type: number, format: double }
        dmi_pct_peso_vivo: { type: number, format: double }
    EstimacionNutricionalResponse:
      allOf:
        - $ref: "#/components/schemas/EstimacionDemanda"
        - type: object
          properties:
            advertencias:
              type: array
              items: { type: string }
              description: Del modelo predictivo, por animal (ej. peso fuera de rango observado para la categoría); no se persisten
            advertencias_generales:
              type: array
              items: { type: string }
              description: Limitaciones fijas del modelo (ej. entrenado con datos de Kenia, no incluye clima/forraje), no dependen del lote de animales; no se persisten
            predicciones:
              type: array
              items: { $ref: "#/components/schemas/DmiPrediccionAnimal" }
              description: Detalle por animal devuelto por el modelo; no se persiste (EstimacionDemanda es agregada por potrero)
...